+++ SD6503_DHLPROJECT/Controllers/TransactionTablesController.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SD6503_DHLPROJECT.Models;
+using Microsoft.AspNetCore.Http;
+
 
 namespace SD6503_DHLPROJECT.Controllers
 {
@@ -21,34 +23,55 @@ namespace SD6503_DHLPROJECT.Controllers
         // GET: TransactionTables
         public async Task<IActionResult> Index()
         {
-            var dhlDatabaseContext = _context.TransactionTables.Include(t => t.FromAccountNavigation);
-            return View(await dhlDatabaseContext.ToListAsync());
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
+            {
+                var dhlDatabaseContext = _context.TransactionTables.Include(t => t.FromAccountNavigation);
+                return View(await dhlDatabaseContext.ToListAsync());
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: TransactionTables/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var transactionTable = await _context.TransactionTables
-                .Include(t => t.FromAccountNavigation)
-                .FirstOrDefaultAsync(m => m.TransactionNumber == id);
-            if (transactionTable == null)
+                var transactionTable = await _context.TransactionTables
+                    .Include(t => t.FromAccountNavigation)
+                    .FirstOrDefaultAsync(m => m.TransactionNumber == id);
+                if (transactionTable == null)
+                {
+                    return NotFound();
+                }
+
+                return View(transactionTable);
+            }
+            else
             {
-                return NotFound();
-            }
-
-            return View(transactionTable);
+                return RedirectToAction("Index", "Home");
+            }          
         }
 
         // GET: TransactionTables/Create
         public IActionResult Create()
         {
-            ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name");
-            return View();
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
+            {
+                ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name");
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // POST: TransactionTables/Create
@@ -58,31 +81,47 @@ namespace SD6503_DHLPROJECT.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("TransactionNumber,ToAccount,FromAccount,LendAmount,PaybackAmount")] TransactionTable transactionTable)
         {
-            if (ModelState.IsValid)
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
             {
-                _context.Add(transactionTable);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(transactionTable);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name", transactionTable.FromAccount);
+                return View(transactionTable);
             }
-            ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name", transactionTable.FromAccount);
-            return View(transactionTable);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+     
         }
 
         // GET: TransactionTables/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
             {
-                return NotFound();
+                if (id == null)
+                {
+                    return NotFound();
+                }
+
+                var transactionTable = await _context.TransactionTables.FindAsync(id);
+                if (transactionTable == null)
+                {
+                    return NotFound();
+                }
+                ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name", transactionTable.FromAccount);
+                return View(transactionTable);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
             }
 
-            var transactionTable = await _context.TransactionTables.FindAsync(id);
-            if (transactionTable == null)
-            {
-                return NotFound();
-            }
-            ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name", transactionTable.FromAccount);
-            return View(transactionTable);
         }
 
         // POST: TransactionTables/Edit/5
@@ -92,52 +131,68 @@ namespace SD6503_DHLPROJECT.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("TransactionNumber,ToAccount,FromAccount,LendAmount,PaybackAmount")] TransactionTable transactionTable)
         {
-            if (id != transactionTable.TransactionNumber)
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
             {
-                return NotFound();
-            }
+                if (id != transactionTable.TransactionNumber)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(transactionTable);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!TransactionTableExists(transactionTable.TransactionNumber))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(transactionTable);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!TransactionTableExists(transactionTable.TransactionNumber))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name", transactionTable.FromAccount);
+                return View(transactionTable);
             }
-            ViewData["FromAccount"] = new SelectList(_context.AccountDetails, "AccountNumber", "Name", transactionTable.FromAccount);
-            return View(transactionTable);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
 
         // GET: TransactionTables/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
+            {
+
+                var transactionTable = await _context.TransactionTables
+                    .Include(t => t.FromAccountNavigation)
+                    .FirstOrDefaultAsync(m => m.TransactionNumber == id);
+                if (transactionTable == null)
+                {
+                    return NotFound();
+                }
+
+                return View(transactionTable);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return NotFound();
             }
 
-            var transactionTable = await _context.TransactionTables
-                .Include(t => t.FromAccountNavigation)
-                .FirstOrDefaultAsync(m => m.TransactionNumber == id);
-            if (transactionTable == null)
-            {
-                return NotFound();
-            }
-
-            return View(transactionTable);
         }
 
         // POST: TransactionTables/Delete/5
@@ -145,15 +200,31 @@ namespace SD6503_DHLPROJECT.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var transactionTable = await _context.TransactionTables.FindAsync(id);
-            _context.TransactionTables.Remove(transactionTable);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
+            {
+                var transactionTable = await _context.TransactionTables.FindAsync(id);
+                _context.TransactionTables.Remove(transactionTable);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
         }
 
         private bool TransactionTableExists(int id)
         {
-            return _context.TransactionTables.Any(e => e.TransactionNumber == id);
+            if (HttpContext.Session.GetString("Identifier") != null && HttpContext.Session.GetString("IsAdmin") == "True")
+            {
+                return _context.TransactionTables.Any(e => e.TransactionNumber == id);
+            }
+            else
+            {
+                return false;
+            }
+
         }
     }
 }
